Extract the comparator to its own class
